Added movable player
TileGameState has been renamed to MapViewGameState. Currently, there is a an @ symbol that can be moved around by pressing the arrow keys on the keyboard. The symbol is not perfectly positioned in its tile, which is a problem. Also, it looks pretty fuzzy and bad. The tiles of the game state also now change randomly every few turns.